requirements.txt
mediapipe == 0.8.3.1
# opencv-python
Pillow == 8.1.2